requirements.txt
absl-py==0.9.0
amqp==2.2.1
asgiref==3.2.10
astor==0.8.1
astroid==1.5.3
attrs==19.3.0
autopep8==1.4.4
Babel==2.5.0
backcall==0.1.0
backports.functools-lru-cache==1.6.1
backports.tempfile==1.0
backports.weakref==1.0.post1
beautifulsoup4==4.9.1
billiard==3.5.0.3
bleach==3.1.5
cairocffi==0.8.0
CairoSVG==2.0.3
celery==4.1.0
certifi==2017.7.27.1
gunicorn
pandas
requests_cache
cffi==1.14.2
chardet==3.0.4
click==7.1.2
Django==2.2
django-allauth==0.39.1
django-autoslug-iplweb==1.9.4
django-bootstrap4==0.0.7
django-countries==5.3.3
django-crispy-forms==1.7.2
django-debug-toolbar==1.10.1
django-paypal==0.4.1
django-pipeline==2.0.5
django-polymorphic==3.0.0
djangorestframework==3.8.2
entrypoints==0.3
filelock==3.0.12
Flask==2.0.1
flower==0.9.2
future==0.18.2